requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
frozenlist==1.8.0
gTTS==2.5.4
google-generativeai==0.8.5
groq==0.22.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
orjson==3.11.7
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
Pillow==11.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.33.0
rich==14.3.3
rich-toolkit==0.19.7
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
twilio==9.0.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
